+++ macros/sql/get_column_values.sql
@@ -10,11 +10,11 @@ Returns:
     A list of distinct values for the specified columns
 #}
 
-{% macro get_column_values(table, column, max_records=none, default=none) -%}
+{% macro get_column_values(table, column, sort_column=none, sort_direction=none, max_records=none, default=none) -%}
     {{ return(adapter.dispatch('get_column_values', packages = dbt_utils._get_utils_namespaces())(table, column, max_records, default)) }}
 {% endmacro %}
 
-{% macro default__get_column_values(table, column, max_records=none, default=none) -%}
+{% macro default__get_column_values(table, column, sort_column=none, sort_direction=none, max_records=none, default=none) -%}
 
 {#-- Prevent querying of db in parsing mode. This works because this macro does not create any new refs. #}
     {%- if not execute -%}
@@ -22,10 +22,21 @@ Returns:
     {% endif %}
 {#--  #}
 
+    {%- set order_by = order_by if order_by else 'count(*)' -%}
+    {%- set order_by = 'max(' ~ order_by ~ ')' if order_by == column else order_by -%}
+    {%- set sort_direction -%}
+        {%- if order_by == column -%}
+        {{ sort_direction or 'asc' }} 
+        {%- else -%}
+        {{ sort_direction or 'desc' }}
+        {%- endif -%}
+    {%- endset -%}
+
     {%- set target_relation = adapter.get_relation(database=table.database,
                                           schema=table.schema,
                                          identifier=table.identifier) -%}
 
+    {# If no sort column is supplied, we use the default descending frequency count. #}
     {%- call statement('get_column_values', fetch_result=true) %}
 
         {%- if not target_relation and default is none -%}
@@ -40,16 +51,27 @@ Returns:
 
         {%- else -%}
 
-            select
-                {{ column }} as value
+            with sorted_column_values as (
 
-            from {{ target_relation }}
-            group by 1
-            order by count(*) desc
+                select
+                    {{ column }} as value,
+                    {%- if sort_column is none %}
+                    count(*) as sort_column
+                    {% else %}
+                    {# We take the max sort value for each value to make sure
+                        there are no duplicate rows for each value #}
+                    max({{ sort_column }}) as sort_column
+                    {% endif %}
+                from {{ target_relation }}
+                group by 1
+                order by {{ order_by }} {{ sort_direction }}
+                {% if max_records is not none %}
 
-            {% if max_records is not none %}
-            limit {{ max_records }}
-            {% endif %}
+                limit {{ max_records }}
+                {% endif %}
+
+            )
+            select value from sorted_column_values
 
         {% endif %}